Move PEResponse class to puppetenterprise.apimanagers
